encapsulated validation behaviour for components in "Validating Component Mixin". Applied to Validating Text Input
